# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Sistem Perwalian Mahasiswa STMIK Bandung",
        "description": "Dokumentasi REST API Enterprise untuk Sistem Perwalian Mahasiswa STMIK Bandung. Mengelola Autentikasi Sanctum, Data Dosen Wali, Data Mahasiswa, Pengajuan dan Review Perwalian, serta Rekapitulasi Laporan.",
        "contact": {
            "name": "Tim Capstone STMIK Bandung",
            "email": "admin@stmikbandung.ac.id"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "Server API Lokal STMIK Bandung"
        }
    ],
    "paths": {
...
        "/perwalian": {
            "get": {
                "tags": [
                    "Perwalian"
                ],
                "summary": "Daftar Perwalian (Filter Role, Status, Semester)",
                "operationId": "64ae2d46383e444262b8526a4c37c2ac",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter Status: Pending/Disetujui/Ditolak",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mendapatkan data perwalian"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Perwalian"
                ],
                "summary": "Pengajuan Perwalian Baru oleh Mahasiswa",
                "operationId": "2e56cdfc838f05f26ee92858d16d94dc",
                "responses": {
                    "201": {
                        "description": "Pengajuan perwalian berhasil disimpan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/perwalian/{id}": {
            "get": {
                "tags": [
                    "Perwalian"
                ],
                "summary": "Detail Data Perwalian & Timeline Histori Audit",
                "operationId": "b2d4a197d62f1fd6cd43fa4ac5b4df87",
                "responses": {
                    "200": {
                        "description": "Detail perwalian ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Perwalian"
                ],
                "summary": "Update Data Perwalian (Hanya untuk Status Pending)",
                "operationId": "d22316431cca2529dcdef052154702f6",
                "responses": {
                    "200": {
                        "description": "Perwalian berhasil diperbarui"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Perwalian"
                ],
                "summary": "Hapus Data Perwalian (Hanya untuk Status Pending)",
                "operationId": "32b09bc5dd3843c67037dd5fdcb21a1f",
                "responses": {
                    "200": {
                        "description": "Perwalian dihapus"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/perwalian/{id}/approve-reject": {
            "post": {
                "tags": [
                    "Perwalian"
                ],
                "summary": "Approval/Rejection Perwalian oleh Dosen Wali",
                "operationId": "e0f439da92cce8173700813105a60b9c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "Disetujui",
                                            "Ditolak"
                                        ],
                                        "example": "Disetujui"
                                    },
                                    "catatan_dosen": {
                                        "type": "string",
                                        "example": "Rencana mata kuliah sudah tepat. Lanjutkan perwalian."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Persetujuan perwalian berhasil diproses"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan token Bearer Sanctum hasil autentikasi login (Contoh: 'Bearer {token}')",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}